requirements.txt
numpy==1.23.2
pandas==1.4.3
psycopg2-binary==2.9.3
PyYAML==6.0
requests==2.28.1
